+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Performances: Changed image decoding to auto and lazy loaded
+- Performances: Preact hover images *should* not be rendered on mobile
+- Performances: Preact Throttled less mouse position to have smoother experience
+- Performances: Preact Used browser API to get requestAF and ensure ticking is correctly done
 
 ## [1.1.25] - 2024-12-25
 ### Changed
@@ -346,7 +351,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Few issues on dark mode and event listener
 
-[unreleased]: https://github.com/RSickenberg/rsickenberg.me/compare/v1.1.25...HEAD
+[Unreleased]: https://github.com/RSickenberg/rsickenberg.me/compare/v1.1.25...HEAD
 [1.1.25]: https://github.com/RSickenberg/rsickenberg.me/compare/v1.1.24...v1.1.25
 [1.1.24]: https://github.com/RSickenberg/rsickenberg.me/compare/v1.1.23...v1.1.24
 [1.1.23]: https://github.com/RSickenberg/rsickenberg.me/compare/v1.1.22...v1.1.23